get_vision_embedding and get_vllm_embedding patches
